backtest

See merge request s03-bigdata-sub3/s03p23a105!53

Former-commit-id: bbf573b23868d0fc23205ac55713894f2197c3cd

+++ backend/backtest/models.py
@@ -7,7 +7,7 @@ class Stock(models.Model):
     code = models.CharField(max_length=6)
     name = models.CharField(max_length=30)
     quantity = models.IntegerField()
-    price = models.FloatField()
+    price = models.IntegerField()
     class Meta:
         abstract = True
 
@@ -20,7 +20,7 @@ class Log(models.Model):
 
 class BacktestDate(models.Model):
     date = models.CharField(max_length=100)
-    budget = models.FloatField()
+    budget = models.IntegerField()
     
     class Meta:
         abstract = True
